requirements.txt
dagster
dagster-webserver
cloudscraper
beautifulsoup4
PyMuPDF
python-docx
python-pptx
chromadb
langchain
langchain-chroma
langchain-groq
langchain-ollama
langchain-google-genai
langchain-community
sentence-transformers
ollama
pandas
python-dotenv
gradio
groq
matplotlib
langgraph
flask
flask-session
gunicorn
